requirements.txt
fastapi
python-multipart
tinydb
uvicorn[standard]
jinja2
tensorflow[and-cuda]
tensorflow-datasets
pillow